requirements.txt
Django==2.2.19
django-rest-framework==3.12.2